modified:   app.py

+++ app.py
@@ -8,11 +8,11 @@ app = Flask(__name__)
 
 def conectar_banco():
     try:
-        conexao = psycopg2.connect (
-            host = 'datalake_menu.postgresql.dbaas.com.br',
-            database = 'datalake_menu',
-            user = 'datalake_menu',
-            password = 'Acesso1!'
+        conexao = psycopg2.connect(
+            host='datalake_menu.postgresql.dbaas.com.br',
+            database='datalake_menu',
+            user='datalake_menu',
+            password='Acesso1!'
         )
         return conexao
     except Exception as e:
@@ -30,7 +30,7 @@ def obter_distancia_banco(endereco1, endereco2):
             if resultado:
                 return resultado[0]
             else:
-                return  None
+                return None
         except Exception as e:
             print(f'Erro ao consultar o banco: {e}')
             return None
@@ -43,18 +43,18 @@ def registrar_distancia_banco(endereco1, endereco2, distancia):
         try:
             cursor = conexao.cursor()
             query = """INSERT INTO distancias (endereco1, endereco2, distancia) VALUES (%s, %s, %s)"""
-            cursor.execute(query, (endereco1, endereco2, distancia))
+            cursor.execute(query, (endereco1, endereco2, float(distancia)))  # Conversão aqui
             conexao.commit()
         except Exception as e:
             print(f"Erro ao registrar no banco: {e}")
         finally:
             conexao.close()
-            
+
 def obter_coordenadas(endereco):
     url = f"https://nominatim.openstreetmap.org/search?q={endereco}&format=json"
     response = requests.get(url, headers={'User-Agent': 'Mozilla/5.0'})
     dados = response.json()
-    
+
     if dados:
         return float(dados[0]['lat']), float(dados[0]['lon'])
     else:
@@ -63,13 +63,13 @@ def obter_coordenadas(endereco):
 def calcular_distancia(endereco1, endereco2):
     coordenadas1 = obter_coordenadas(endereco1)
     coordenadas2 = obter_coordenadas(endereco2)
-    
+
     if coordenadas1 is None or coordenadas2 is None:
         print("Não foi possível obter as coordenadas.")
         return None
-    
+
     distancia = geodesic(coordenadas1, coordenadas2).kilometers
-    return np.round(distancia,2)
+    return float(np.round(distancia, 2))  # Conversão aqui também
 
 def classificar_area(distancia):
     areas_de_entrega = {
@@ -85,24 +85,23 @@ def classificar_area(distancia):
         "J": (9, 9.99),
         "K": (10, 10.99),
     }
-    
+
     for area, (min_km, max_km) in areas_de_entrega.items():
         if min_km <= distancia < max_km:
             return area
-    
+
     return "Fora da área de entrega"
 
 def calcular(endereco1, endereco2):
-
-
     distancia = obter_distancia_banco(endereco1, endereco2)
-    
+
     if distancia is None:
         distancia = calcular_distancia(endereco1, endereco2)
         if distancia is not None:
             registrar_distancia_banco(endereco1, endereco2, distancia)
         else:
-           return None, "Não foi possível calcular a distância."
+            return None, "Não foi possível calcular a distância."
+
     area = classificar_area(distancia)
     return distancia, area
 
@@ -116,7 +115,7 @@ def index():
 
         if distancia is None:
             return render_template("index.html", erro=area_ou_erro)
-        
+
         return render_template("index.html", distancia=distancia, area=area_ou_erro)
     return render_template("index.html")
